+++ src/circuit/ops/layouts.rs
@@ -724,6 +724,14 @@ pub fn iff<F: PrimeField + TensorType + PartialOrd>(
     // if mask > 0 then output a else output b
     let (mask, b, a) = (&values[0], &values[1], &values[2]);
 
+    let mask = nonlinearity(
+        config,
+        region.clone(),
+        &[mask.clone()],
+        &LookupOp::GreaterThan { a: 0.0.into() },
+        offset,
+    )?;
+
     let mut lock = region.lock().unwrap();
     let unit: ValTensor<F> = if let Some(region) = lock.as_mut() {
         Tensor::from(
@@ -737,7 +745,7 @@ pub fn iff<F: PrimeField + TensorType + PartialOrd>(
     *offset += 1;
 
     // make sure mask is boolean
-    let assigned_mask = config.inputs[1].assign(&mut lock, *offset, mask)?;
+    let assigned_mask = config.inputs[1].assign(&mut lock, *offset, &mask)?;
     if let Some(region) = lock.as_mut() {
         for i in 0..assigned_mask.len() {
             let (x, y) = config.inputs[1].cartesian_coord(*offset + i);
