+++ src/circuit/ops/poly.rs
@@ -1,12 +1,8 @@
-use std::sync::{Arc, Mutex};
-
-use itertools::Itertools;
-
 use crate::{
     circuit::layouts,
-    graph::scale_to_multiplier,
     tensor::{self, Tensor, TensorError},
 };
+use std::sync::{Arc, Mutex};
 
 use super::{base::BaseOp, *};
 
@@ -49,7 +45,6 @@ pub enum PolyOp<F: PrimeField + TensorType + PartialOrd> {
     Pow(u32),
     Pack(u32, u32),
     GlobalSumPool,
-    Iff,
     Concat {
         axis: usize,
     },
@@ -60,46 +55,7 @@ pub enum PolyOp<F: PrimeField + TensorType + PartialOrd> {
     },
 }
 
-impl<F: PrimeField + TensorType + PartialOrd> PolyOp<F> {
-    fn homogenize_input_scales(
-        &self,
-        input_scales: Vec<u32>,
-        inputs_to_scale: Vec<usize>,
-    ) -> Result<Box<dyn Op<F>>, Box<dyn Error>> {
-        if inputs_to_scale.is_empty() {
-            return Ok(Box::new(self.clone()));
-        }
-
-        let mut dividers: Vec<u128> = vec![1; input_scales.len()];
-        if !input_scales.windows(2).all(|w| w[0] == w[1]) {
-            let min_scale = input_scales.iter().min().unwrap();
-            let _ = input_scales
-                .iter()
-                .enumerate()
-                .map(|(idx, input_scale)| {
-                    if !inputs_to_scale.contains(&idx) {
-                        return;
-                    }
-                    let scale_diff = input_scale - min_scale;
-                    if scale_diff > 0 {
-                        let mult = scale_to_multiplier(scale_diff);
-                        dividers[idx] = mult as u128;
-                    }
-                })
-                .collect_vec();
-        }
-
-        // only rescale if need to
-        if dividers.iter().any(|&x| x > 1) {
-            Ok(Box::new(crate::circuit::Rescaled {
-                inner: Box::new(self.clone()),
-                scale: (0..input_scales.len()).zip(dividers).collect_vec(),
-            }))
-        } else {
-            Ok(Box::new(self.clone()))
-        }
-    }
-}
+impl<F: PrimeField + TensorType + PartialOrd> PolyOp<F> {}
 
 impl<F: PrimeField + TensorType + PartialOrd> Op<F> for PolyOp<F> {
     fn as_any(&self) -> &dyn Any {
@@ -121,7 +77,6 @@ impl<F: PrimeField + TensorType + PartialOrd> Op<F> for PolyOp<F> {
             PolyOp::GlobalSumPool => "GLOBALSUMPOOL",
             PolyOp::Conv { .. } => "CONV",
             PolyOp::SumPool { .. } => "SUMPOOL",
-            PolyOp::Iff => "IFF",
             PolyOp::Gather { .. } => "GATHER",
             PolyOp::Concat { .. } => "CONCAT",
             PolyOp::Slice { .. } => "SLICE",
@@ -201,17 +156,6 @@ impl<F: PrimeField + TensorType + PartialOrd> Op<F> for PolyOp<F> {
                 tensor::ops::sum_axes(&inputs[0], axes)
             }
             PolyOp::GlobalSumPool => unreachable!(),
-            PolyOp::Iff => {
-                let mask = inputs[0].clone();
-                // if mask > 0 then output a else output b
-                let a = inputs[2].clone();
-                let b = inputs[1].clone();
-
-                let masked_a = (mask.clone() * a)?;
-                let masked_b = ((Tensor::from(vec![1_i128].into_iter()) - mask)? * b)?;
-
-                masked_a + masked_b
-            }
             PolyOp::Concat { axis } => {
                 if inputs.len() < 2 {
                     return Err(TensorError::DimMismatch("concat inputs".to_string()));
@@ -244,7 +188,6 @@ impl<F: PrimeField + TensorType + PartialOrd> Op<F> for PolyOp<F> {
             PolyOp::Gather { dim, index } => {
                 tensor::ops::gather(&values[0].get_inner_tensor()?, *dim, index)?.into()
             }
-            PolyOp::Iff => layouts::iff(config, region, values[..].try_into()?, offset)?,
             PolyOp::Sum { axes } => {
                 layouts::sum_axes(config, region, values[..].try_into()?, axes, offset)?
             }
@@ -344,7 +287,7 @@ impl<F: PrimeField + TensorType + PartialOrd> Op<F> for PolyOp<F> {
                 scale
             }
             PolyOp::Gather { .. } => in_scales[0],
-            PolyOp::Iff => in_scales[1],
+
             PolyOp::Sum { .. } => in_scales[0],
             PolyOp::Conv { kernel, bias, .. } => {
                 let output_scale = in_scales[0] + kernel.scale();
@@ -389,8 +332,7 @@ impl<F: PrimeField + TensorType + PartialOrd> Op<F> for PolyOp<F> {
     fn rescale(&self, input_scales: Vec<u32>, _: u32) -> Box<dyn Op<F>> {
         let inputs_to_scale = self.requires_homogenous_input_scales();
         // creates a rescaled op if the inputs are not homogenous
-        self.homogenize_input_scales(input_scales, inputs_to_scale)
-            .unwrap()
+        homogenize_input_scales::<F>(self.clone(), input_scales, inputs_to_scale).unwrap()
     }
 
     fn requires_homogenous_input_scales(&self) -> Vec<usize> {
@@ -399,8 +341,6 @@ impl<F: PrimeField + TensorType + PartialOrd> Op<F> for PolyOp<F> {
             PolyOp::Add { .. } | PolyOp::Sub | PolyOp::Mult { .. } | PolyOp::Einsum { .. }
         ) {
             vec![0, 1]
-        } else if matches!(self, PolyOp::Iff) {
-            vec![1, 2]
         } else {
             vec![]
         }
